fixup! 15153: Implement new events pattern
